fastapi-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytest
pytest-asyncio
httpx
openai
python-dotenv
faiss-cpu
sentence-transformers
beautifulsoup4
requests
sqlite3
sqlalchemy
chromadb
numpy
pandas